fix bug add userWithTags
