+++ server/src/tag/tag.entity.ts
@@ -1,7 +1,5 @@
-import { Entity, PrimaryGeneratedColumn, Column, ManyToMany } from 'typeorm';
+import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
 import { User } from '../user/user.entity';
-
-
 @Entity()
 export class Tag {
     @PrimaryGeneratedColumn()
