+++ server/src/user/user.entity.ts
@@ -12,15 +12,11 @@ export class User {
   @Column({ select: false })
   password!: string;
 
-  @ManyToMany(() => Tag, (tag) => tag.users)
-  @JoinTable()
-  tags!: Tag[];
+   @ManyToMany(() => Tag)
+    @JoinTable()
+    technologies!: Tag[]
 
-  @ManyToMany(() => Tag, tag => tag.specializedUsers)
-  @JoinTable()
-  specializations!: Tag[];
-
-  @ManyToMany(() => Tag, tag => tag.technologies)
-  @JoinTable()
-  technologies!: Tag[];
+    @ManyToMany(() => Tag)
+    @JoinTable()
+    specializations!: Tag[]
 }
